ELEC-203: Moved TX to CAN FSM - untested
